pyproject.toml
[project]
name = "aftermap"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.12.5",
    "pyzmq>=27.1.0",
    "sqlalchemy>=2.0.49",
]
